refactor cd structure
