+++ cmd/plural/cd_providers.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/pkg/errors"
 	gqlclient "github.com/pluralsh/console-client-go"
 	"github.com/pluralsh/plural/pkg/api"
 	"github.com/pluralsh/plural/pkg/utils"
@@ -17,48 +16,22 @@ import (
 
 const kindSecret = "Secret"
 
-func (p *Plural) handleDeleteProviderCredentials(c *cli.Context) error {
-	id := c.Args().Get(0)
-	if err := p.InitConsoleClient(consoleToken, consoleURL); err != nil {
-		return err
+func (p *Plural) cdProviders() cli.Command {
+	return cli.Command{
+		Name:        "providers",
+		Subcommands: p.cdProvidersCommands(),
+		Usage:       "manage CD providers",
 	}
-	if _, err := p.ConsoleClient.DeleteProviderCredentials(id); err != nil {
-		return err
-	}
-	utils.Success("Provider credential %s has been deleted successfully", id)
-	return nil
 }
 
-func (p *Plural) handleCreateProviderCredentials(c *cli.Context) error {
-	providerName := c.Args().Get(0)
-	if err := p.InitConsoleClient(consoleToken, consoleURL); err != nil {
-		return err
+func (p *Plural) cdProvidersCommands() []cli.Command {
+	return []cli.Command{
+		{
+			Name:   "list",
+			Action: latestVersion(p.handleListProviders),
+			Usage:  "list providers",
+		},
 	}
-
-	attr, err := p.credentialsPreflights()
-	if err != nil {
-		return err
-	}
-
-	resp, err := p.ConsoleClient.CreateProviderCredentials(providerName, *attr)
-	if err != nil {
-		errList := errors.New("CreateProviderCredentials")
-		errList = errors.Wrap(errList, err.Error())
-		if *attr.Kind == kindSecret {
-			if err := p.SecretDelete(*attr.Namespace, attr.Name); err != nil {
-				errList = errors.Wrap(errList, err.Error())
-			}
-		}
-		return errList
-	}
-	if resp == nil {
-		return fmt.Errorf("the response from CreateProviderCredentials is empty")
-	}
-
-	headers := []string{"Id", "Name", "Namespace"}
-	return utils.PrintTable([]*gqlclient.ProviderCredentialFragment{resp.CreateProviderCredential}, headers, func(sd *gqlclient.ProviderCredentialFragment) ([]string, error) {
-		return []string{sd.ID, sd.Name, sd.Namespace}, nil
-	})
 }
 
 func (p *Plural) handleListProviders(_ *cli.Context) error {
